.
开始getbyid需要的参数是long，传入long之后报错说传入的integer。更改basedao的参数类型为int，运行之后原来正常的getbyid（）0报错，变成传入int变成了Long，之前报错的getbyid（）正确了。

+++ src/com/zm/dao/IBaseDao.java
@@ -6,13 +6,13 @@ public interface IBaseDao<T> {
 	
 	public void add(T entity);
 
-	public void delet(Long id);
+	public void delet(long id);
 
 	public void update(T entity);
 
 	public List<T> findall();
 
-	public T getById(Long id);
+	public T getById(int id);
 
-	public List<T> getByIds(Long[] ids);
+	public List<T> getByIds(long[] ids);
 }
